+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## Unreleased
 
 - Fixed an issue where new commits wouldn't trigger updates. [#951](https://github.com/pulumi/pulumi-kubernetes-operator/issues/951)
+- New feature: Dynamic environment variables [#976](https://github.com/pulumi/pulumi-kubernetes-operator/pull/976)
 
 ## 2.1.0 (2025-05-29)